Wrote 2-storage-values
