+++ pallets/single-value/src/lib.rs
@@ -1,248 +1,38 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 /// Single Value Storage
-use support::{decl_module, decl_event, decl_storage, ensure, dispatch::DispatchResult, StorageValue};
-use system::{self, ensure_signed};
-
-pub trait Trait: system::Trait {
-    type Event: From<Event<Self>> + Into<<Self as system::Trait>::Event>;
-}
-
-decl_storage! {
-    trait Store for Module<T: Trait> as SingleValue {
-        MyValue: u32;
-        MyAccount: T::AccountId;
-    }
-}
-
-decl_event!(
-    pub enum Event<T>
-    where
-        AccountId = <T as system::Trait>::AccountId,
-        BlockNumber = <T as system::Trait>::BlockNumber,
-    {
-        // value set at Set, BlockNumber at time
-        ValueSet(u32, BlockNumber),
-        // requested public value get
-        ValueGet(u32, BlockNumber),
-        // requested public account get
-        AccountGet(AccountId, BlockNumber),
-        // account set at Set, BlockNumber at time
-        AccountSet(AccountId, BlockNumber),
-    }
-);
-
-decl_module! {
-    pub struct Module<T: Trait> for enum Call where origin: T::Origin {
-        fn deposit_event() = default;
-
-        fn set_value(origin, value: u32) -> DispatchResult {
-            let _ = ensure_signed(origin)?;
-            let now = <system::Module<T>>::block_number();
-            <MyValue>::put(value.clone());
-            Self::deposit_event(RawEvent::ValueSet(value, now));
-            Ok(())
-        }
-
-        fn get_value(origin) -> DispatchResult {
-            let _ = ensure_signed(origin)?;
-            let now = <system::Module<T>>::block_number();
-            ensure!(<MyValue>::exists(), "value does not exist");
-            let val = <MyValue>::get();
-            Self::deposit_event(RawEvent::ValueGet(val, now));
-            Ok(())
-        }
-
-        fn set_account(origin, account_to_set: T::AccountId) -> DispatchResult {
-            let _ = ensure_signed(origin)?;
-            let now = <system::Module<T>>::block_number();
-            <MyAccount<T>>::put(account_to_set.clone());
-            Self::deposit_event(RawEvent::AccountSet(account_to_set, now));
-            Ok(())
-        }
-
-        fn get_account(origin) -> DispatchResult {
-            let _ = ensure_signed(origin)?;
-            let now = <system::Module<T>>::block_number();
-            ensure!(<MyAccount<T>>::exists(), "account dne");
-            let got_account = <MyAccount<T>>::get();
-            Self::deposit_event(RawEvent::AccountGet(got_account, now));
-            Ok(())
-        }
-    }
-}
+use frame_support::{decl_module, decl_storage, dispatch::DispatchResult};
+use frame_system::{self as system, ensure_signed};
 
 #[cfg(test)]
-mod tests {
-	use support::{assert_err, assert_ok, impl_outer_origin, impl_outer_event, parameter_types};
-	use runtime_primitives::{Perbill, traits::{IdentityLookup, BlakeTwo256}, testing::Header};
-	use super::RawEvent;
-	use runtime_io;
-	use primitives::H256;
-	use crate::{Module, Trait};
+mod tests;
 
-	impl_outer_origin!{
-		pub enum Origin for TestRuntime {}
+pub trait Trait: system::Trait {}
+
+decl_storage! {
+	trait Store for Module<T: Trait> as SingleValue {
+		StoredValue get(fn stored_value): u32;
+		StoredAccount get(fn stored_account): T::AccountId;
 	}
+}
 
-	// Workaround for https://github.com/rust-lang/rust/issues/26925 . Remove when sorted.
-	#[derive(Clone, PartialEq, Eq, Debug)]
-	pub struct TestRuntime;
-	parameter_types! {
-		pub const BlockHashCount: u64 = 250;
-		pub const MaximumBlockWeight: u32 = 1024;
-		pub const MaximumBlockLength: u32 = 2 * 1024;
-		pub const AvailableBlockRatio: Perbill = Perbill::one();
-	}
-	impl system::Trait for TestRuntime {
-		type Origin = Origin;
-		type Index = u64;
-		type Call = ();
-		type BlockNumber = u64;
-		type Hash = H256;
-		type Hashing = BlakeTwo256;
-		type AccountId = u64;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type Header = Header;
-		type Event = TestEvent;
-		type BlockHashCount = BlockHashCount;
-		type MaximumBlockWeight = MaximumBlockWeight;
-		type MaximumBlockLength = MaximumBlockLength;
-		type AvailableBlockRatio = AvailableBlockRatio;
-		type Version = ();
-		type ModuleToIndex = ();
-	}
+decl_module! {
+	pub struct Module<T: Trait> for enum Call where origin: T::Origin {
 
-    mod single_value {
-        pub use crate::Event;
-    }
+		fn set_value(origin, value: u32) -> DispatchResult {
+			let _ = ensure_signed(origin)?;
 
-    impl_outer_event! {
-        pub enum TestEvent for TestRuntime {
-            single_value<T>,
-        }
-    }
+			StoredValue::put(value);
 
-	impl Trait for TestRuntime {
-		type Event = TestEvent;
-	}
+			Ok(())
+		}
 
-	pub type System = system::Module<TestRuntime>;
-	pub type SingleValue = Module<TestRuntime>;
+		fn set_account(origin) -> DispatchResult {
+			let who = ensure_signed(origin)?;
 
-	pub struct ExtBuilder;
+			<StoredAccount<T>>::put(&who);
 
-	impl ExtBuilder {
-		pub fn build() -> runtime_io::TestExternalities {
-			let storage = system::GenesisConfig::default().build_storage::<TestRuntime>().unwrap();
-			runtime_io::TestExternalities::from(storage)
+			Ok(())
 		}
 	}
-
-    #[test]
-    fn set_value_works() {
-        ExtBuilder::build().execute_with(|| {
-            System::set_block_number(2);
-            assert_ok!(SingleValue::set_value(Origin::signed(1), 10));
-
-            let expected_event = TestEvent::single_value(
-                RawEvent::ValueSet(10, 2),
-            );
-
-            assert!(System::events().iter().any(|a| a.event == expected_event));
-
-            System::set_block_number(15);
-            assert_ok!(SingleValue::set_value(Origin::signed(1), 11));
-
-            let expected_event = TestEvent::single_value(
-                RawEvent::ValueSet(11, 15),
-            );
-
-			assert!(System::events().iter().any(|a| a.event == expected_event));
-        })
-    }
-
-    #[test]
-    fn set_account_works() {
-        // NOTE: could probably be combined into `set_works()`
-        ExtBuilder::build().execute_with(|| {
-            System::set_block_number(2);
-            assert_ok!(SingleValue::set_account(Origin::signed(1), 10));
-
-            let expected_event = TestEvent::single_value(
-                RawEvent::AccountSet(10, 2),
-            );
-
-            assert!(System::events().iter().any(|a| a.event == expected_event));
-
-            System::set_block_number(15);
-            assert_ok!(SingleValue::set_account(Origin::signed(1), 11));
-
-            let expected_event = TestEvent::single_value(
-                RawEvent::AccountSet(11, 15),
-            );
-
-			assert!(System::events().iter().any(|a| a.event == expected_event));
-        })
-    }
-
-    #[test]
-    fn get_works() {
-        ExtBuilder::build().execute_with(|| {
-
-                assert_err!(
-                    SingleValue::get_value(Origin::signed(1)),
-                    "value does not exist"
-                );
-                assert_err!(
-                    SingleValue::get_account(Origin::signed(2)),
-                    "account dne"
-                );
-
-                // set value and account
-                System::set_block_number(2);
-                assert_ok!(SingleValue::set_value(Origin::signed(2), 5));
-                assert_ok!(SingleValue::set_account(Origin::signed(1), 10));
-
-                // get value and account
-                assert_ok!(SingleValue::get_value(Origin::signed(1)));
-
-                let expected_event = TestEvent::single_value(
-                    RawEvent::ValueGet(5, 2),
-                );
-
-                assert!(System::events().iter().any(|a| a.event == expected_event));
-
-                assert_ok!(SingleValue::get_account(Origin::signed(1)));
-
-                let expected_event2 = TestEvent::single_value(
-                    RawEvent::AccountGet(10, 2),
-                );
-
-                assert!(System::events().iter().any(|a| a.event == expected_event2));
-
-                // reset value and account
-                System::set_block_number(12);
-                assert_ok!(SingleValue::set_value(Origin::signed(2), 27));
-                assert_ok!(SingleValue::set_account(Origin::signed(1), 13));
-
-                // reget value and account
-                assert_ok!(SingleValue::get_value(Origin::signed(1)));
-
-                let expected_event3 = TestEvent::single_value(
-                    RawEvent::ValueGet(27, 12),
-                );
-
-                assert!(System::events().iter().any(|a| a.event == expected_event3));
-
-                assert_ok!(SingleValue::get_account(Origin::signed(1)));
-
-                let expected_event4 = TestEvent::single_value(
-                    RawEvent::AccountGet(13, 12),
-                );
-
-                assert!(System::events().iter().any(|a| a.event == expected_event4));
-            }
-        )
-    }
 }
