+++ text/SUMMARY.md
@@ -6,10 +6,10 @@
     - [Interacting with a Node](./prepare-kitchen/2-interact-node.md)
     - [Kitchen Organization](./prepare-kitchen/3-kitchen-organization.md)
 - [Appetizers](./appetizers/README.md)
+    - [Hello Substrate](./appetizers/1-hello-substrate.md)
+    - [Single Value Storage](./appetizers/2-storage-values.md)
     - [Events Verify Execution](./appetizers/events.md)
     - [Adding Machine](./appetizers/adder.md)
-    - [Single Value Storage](./appetizers/value.md)
-    - [Hello Substrate](./appetizers/hello-substrate.md)
 - [Runtime Storage API](./storage/README.md)
     - [Cache Locally > Storage Calls](./storage/cache.md)
     - [Sets](./storage/iterate.md)
